final fixes to the gifting feature
