+++ messages/en.json
@@ -641,10 +641,10 @@
     "giftTo": "Gift a ticket to {userName}",
     "giftDesc": "By clicking on the button below, you will send a free ticket to {userName}",
     "sent": "Gift Ticket sent!",
-    "sentBody": "Your ticket has successfully been gifted to ${userName}",
+    "sentBody": "Your ticket has successfully been gifted to {userName}",
     "ok": "ok",
     "error": "Unable to send gift Ticket!",
-    "errorBody": "An error occured while sending the gift ticket to ${userName}",
+    "errorBody": "An error occured while sending the gift ticket to {userName}",
     "customInvite": "Create a custom invitation for {searchTerm}",
     "customTitle": "You need to create an account for {searchTerm} first",
     "createAccount": "Create an account for {searchTerm}",
