+++ messages/es.json
@@ -637,10 +637,10 @@
     "giftTo": "Regalar una entrada a {userName}",
     "giftDesc": "Al hacer clic en el botón de abajo, enviarás una entrada gratuita a {userName}",
     "sent": "¡Entrada de Regalo enviada!",
-    "sentBody": "Tu entrada ha sido enviada exitosamente a ${userName}",
+    "sentBody": "Tu entrada ha sido enviada exitosamente a {userName}",
     "ok": "Aceptar",
     "error": "¡No se pudo enviar la entrada de regalo!",
-    "errorBody": "Ocurrió un error al enviar la entrada de regalo a ${userName}",
+    "errorBody": "Ocurrió un error al enviar la entrada de regalo a {userName}",
     "customInvite": "Crear una invitación personalizada para {searchTerm}",
     "customTitle": "Debes crear una cuenta para {searchTerm} primero",
     "createAccount": "Crear una cuenta para {searchTerm}",
