+++ messages/fr.json
@@ -606,10 +606,10 @@
     "giftTo": "Offrir un billet à {userName}",
     "giftDesc": "En cliquant sur le bouton ci-dessous, vous enverrez un billet gratuit à {userName}",
     "sent": "Billet Cadeau envoyé !",
-    "sentBody": "Votre billet a été envoyé avec succès à ${userName}",
+    "sentBody": "Votre billet a été envoyé avec succès à {userName}",
     "ok": "OK",
     "error": "Impossible d'envoyer le Billet Cadeau !",
-    "errorBody": "Une erreur s'est produite lors de l'envoi du billet cadeau à ${userName}",
+    "errorBody": "Une erreur s'est produite lors de l'envoi du billet cadeau à {userName}",
     "customInvite": "Créer une invitation personnalisée pour {searchTerm}",
     "customTitle": "Vous devez d'abord créer un compte pour {searchTerm}",
     "createAccount": "Créer un compte pour {searchTerm}",
