+++ messages/pt.json
@@ -606,10 +606,10 @@
     "giftTo": "Oferecer um bilhete a {userName}",
     "giftDesc": "Ao clicar no botão abaixo, vais enviar um bilhete gratuito a {userName}",
     "sent": "Bilhete de Oferta enviado!",
-    "sentBody": "O teu bilhete foi enviado com sucesso para ${userName}",
+    "sentBody": "O teu bilhete foi enviado com sucesso para {userName}",
     "ok": "OK",
     "error": "Não foi possível enviar o bilhete de oferta!",
-    "errorBody": "Ocorreu um erro ao enviar o bilhete de oferta para ${userName}",
+    "errorBody": "Ocorreu um erro ao enviar o bilhete de oferta para {userName}",
     "customInvite": "Criar um convite personalizado para {searchTerm}",
     "customTitle": "É necessário criar uma conta para {searchTerm} primeiro",
     "createAccount": "Criar uma conta para {searchTerm}",
