+++ src/app/(protected)/admin/organizations/[id]/events/[eventId]/page.tsx
@@ -1,4 +1,3 @@
-import { getEvent } from '@/services/actions/event/getEvent'
 import { getUser } from '@/lib/auth/lucia'
 import OrganizationLayout from '@/components/orgs/OrganizationLayout'
 import { redirect } from 'next/navigation'
@@ -10,13 +9,14 @@ import { getTranslations } from 'next-intl/server'
 import { ArrowLeft, Calendar, Link2, MapPin, Pencil, Trash } from 'lucide-react'
 import { MdPreview } from '@/components/events/editor/MdPreview'
 import { Card, CardContent } from '@/components/ui/card'
+import { getEventById } from '@/services/actions/event/getEventById'
 
 export default async function EventDetailsPage({
   params
 }: {
   params: { id: string; eventId: string }
 }) {
-  const event = await getEvent({ eventId: params.eventId })
+  const event = await getEventById(params.eventId)
   const user = await getUser()
   const attendees = await getEventRegistrations(params.eventId)
