+++ src/services/actions/event/getEvent.ts
@@ -1,29 +1,50 @@
+// 'use server'
+
+// import prisma from '@/lib/prisma'
+// import { SavedEventSchema } from '@/schemas/eventSchema'
+
+// export async function getEvent({
+//   eventId,
+//   includeTickets
+// }: {
+//   eventId: string
+//   includeTickets?: boolean
+// }) {
+//   const event = await prisma.event.findUniqueOrThrow({
+//     where: { id: eventId },
+//     include: { tickets: includeTickets || false }
+//   })
+
+//   const formattedEvent = {
+//     ...event,
+//     tickets: event.tickets?.map(ticket => ({
+//       ...ticket,
+//       price: ticket.price.toNumber() // Convert Decimal to number
+//     }))
+//   }
+
+//   return SavedEventSchema.parse(formattedEvent)
+// }
+
+// export type EventType = Awaited<ReturnType<typeof getEvent>>
+
 'use server'
 
+import { SavedEventType } from '@/schemas/eventSchema'
 import prisma from '@/lib/prisma'
-import { SavedEventSchema } from '@/schemas/eventSchema'
+import {
+  SavedEventSchema,
+  GetEventFilterSchemaType
+} from '@/schemas/eventSchema'
 
 export async function getEvent({
   eventId,
   includeTickets
-}: {
-  eventId: string
-  includeTickets?: boolean
-}) {
+}: GetEventFilterSchemaType): Promise<SavedEventType> {
   const event = await prisma.event.findUniqueOrThrow({
     where: { id: eventId },
-    include: { tickets: includeTickets || false }
+    include: includeTickets ? { tickets: true } : undefined
   })
 
-  const formattedEvent = {
-    ...event,
-    tickets: event.tickets?.map(ticket => ({
-      ...ticket,
-      price: ticket.price.toNumber() // Convert Decimal to number
-    }))
-  }
-
-  return SavedEventSchema.parse(formattedEvent)
+  return SavedEventSchema.parse(event)
 }
-
-export type EventType = Awaited<ReturnType<typeof getEvent>>
